requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.4
crispy-tailwind==1.0.3
cron-descriptor==1.4.5
distlib==0.3.8
Django==5.1.1
django-celery-beat==2.7.0
django-ckeditor-5==0.2.17
django-crispy-forms==2.3
django-environ==0.12.0
django-timezone-field==7.1
django_celery_results==2.6.0
filelock==3.16.0
gunicorn==23.0.0
kombu==5.5.2
packaging==24.2
pillow==11.1.0
platformdirs==4.3.3
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.1
tzdata==2025.2
vine==5.1.0
virtualenv==20.26.4
wcwidth==0.2.13
